+++ Source/EventFlow.Tests/IntegrationTests/EventStores/FilesEventStoreTests.cs
@@ -44,7 +44,7 @@ namespace EventFlow.Tests.IntegrationTests.EventStores
             private IReadModelPopulator _readModelPopulator;
             private IQueryProcessor _queryProcessor;
 
-            public override IRootResolver CreateRootResolver(EventFlowOptions eventFlowOptions)
+            public override IRootResolver CreateRootResolver(IEventFlowOptions eventFlowOptions)
             {
                 var storePath = Path.Combine(
                     Path.GetTempPath(),