+++ Source/EventFlow/Commands/DistinctCommand.cs
@@ -24,18 +24,19 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
+using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Core;
 
 namespace EventFlow.Commands
 {
-    public abstract class DistinctCommand<TAggregate, TIdentity> : ICommand<TAggregate, TIdentity>
+    public abstract class DistinctCommand<TAggregate, TIdentity> : ICommand<TAggregate, TIdentity, CommandId>
         where TAggregate : IAggregateRoot<TIdentity>
         where TIdentity : IIdentity
     {
-        private readonly Lazy<ISourceId> _lazySourceId;
+        private readonly Lazy<CommandId> _lazySourceId;
 
-        public ISourceId SourceId => _lazySourceId.Value;
+        public CommandId SourceId => _lazySourceId.Value;
         public TIdentity AggregateId { get; }
 
         protected DistinctCommand(
@@ -43,12 +44,12 @@ namespace EventFlow.Commands
         {
             if (aggregateId == null) throw new ArgumentNullException(nameof(aggregateId));
 
-            _lazySourceId = new Lazy<ISourceId>(CalculateSourceId, LazyThreadSafetyMode.PublicationOnly);
+            _lazySourceId = new Lazy<CommandId>(CalculateSourceId, LazyThreadSafetyMode.PublicationOnly);
 
             AggregateId = aggregateId;
         }
 
-        private ISourceId CalculateSourceId()
+        private CommandId CalculateSourceId()
         {
             var bytes = GetSourceIdComponents().SelectMany(b => b).ToArray();
             return CommandId.NewDeterministic(
@@ -57,5 +58,15 @@ namespace EventFlow.Commands
         }
 
         protected abstract IEnumerable<byte[]> GetSourceIdComponents();
+
+        public Task<ISourceId> PublishAsync(ICommandBus commandBus, CancellationToken cancellationToken)
+        {
+            return commandBus.PublishAsync(this, cancellationToken);
+        }
+
+        public ISourceId GetSourceId()
+        {
+            return SourceId;
+        }
     }
 }