+++ Source/EventFlow/Commands/ICommandHandler.cs
@@ -31,11 +31,19 @@ namespace EventFlow.Commands
     {
     }
 
-    public interface ICommandHandler<in TAggregate, TIdentity, in TCommand> : ICommandHandler
-        where TAggregate : IAggregateRoot<TIdentity>
-        where TIdentity : IIdentity
-        where TCommand : ICommand<TAggregate, TIdentity>
+    public interface ICommandHandler<in TAggregate, TIdentity, TSourceIdentity, in TCommand> : ICommandHandler
+    where TAggregate : IAggregateRoot<TIdentity>
+    where TIdentity : IIdentity
+    where TSourceIdentity : ISourceId
+    where TCommand : ICommand<TAggregate, TIdentity, TSourceIdentity>
     {
         Task ExecuteAsync(TAggregate aggregate, TCommand command, CancellationToken cancellationToken);
+    }
+
+    public interface ICommandHandler<in TAggregate, TIdentity, in TCommand> : ICommandHandler<TAggregate, TIdentity, ISourceId, TCommand>
+        where TAggregate : IAggregateRoot<TIdentity>
+        where TIdentity : IIdentity
+        where TCommand : ICommand<TAggregate, TIdentity, ISourceId>
+    {
     }
 }